refactor: simply advertise with a go compatible repsonse every 10s
License: MIT
